pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sales-lead-builder"
version = "0.1.0"
description = "CLI tool to enrich sales prospect spreadsheets from Google Sheets"
authors = [
    { name = "Your Team" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.142.0",
    "google-auth>=2.35.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "openai>=1.50.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "tldextract>=5.1.2",
    "phonenumbers>=8.13.46",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.5"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-mock>=3.14.0",
    "responses>=0.25.3"
]

[project.scripts]
sales-lead-builder = "sales_lead_builder.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
